Move psc-publish code into a library
